+++ packages/termunicode/CHANGELOG.md
@@ -4,6 +4,7 @@
 - feature: refactor generator and added smoke tests
 - feature: improve tests for ucd parsers
 - chore: update unicode data to 16.0.0
+- fix: correct width for some ambiguous characters
 
 ## 0.2.0